+++ src/components/buttons/CheckboxButton/CheckboxButton.tsx
@@ -22,7 +22,6 @@ export const CheckboxButton: FC<CheckboxButtonProps> = props => (
 )
 
 export interface CheckboxButtonProps {
-  children?: string
   value: string
   isDisabled?: boolean
   isChecked?: boolean