+++ src/components/contexts/I18nContext.ts
@@ -1,7 +1,7 @@
 import { createContext } from 'react'
 
 type I18nUtils = {
-  t: (key: string, variables: { [errorKey: string]: any }) => string
+  t: (key: string, variables?: { [errorKey: string]: any }) => string
   addTranslations: (language: string, translations: { [key: string]: string }) => void
 }