+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project **don't** adheres to [Semantic Versioning](http://semver.org/sp
 - `ThemeProvider` become a wrapper of the `ThemeContext.Provider`.
   It provides a new way to set theme's value (the `theme` attribute).
   User doesn't need to explicitly mention every theme options.
+- `ThemeProvider` automatically load the font on the whole page.
 
 ## [v0.0.24] - 2019-12-31