+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project **doesn't** adhere to [Semantic Versioning](http://semver.org/s
 - All form components can be updated using the `value` attribute.
 - **BREACKING CHANGES** The default behavior of `TextFieldList` is to be "not ordered".
   You'll have to add the new `isOrdered` attribute on your existing components.
+- `RadioButton` and `CheckboxButton` support other type than string for their `chidren` attribute.
 
 ### Fixed
 - Update types to allow no theme data into `ThemeProvider`.